test: use smaller network for tests

+++ tests/test_clip.py
@@ -23,7 +23,7 @@ import torchopt
 
 
 @helpers.parametrize(
-    dtype=[torch.float32, torch.float64],
+    dtype=[torch.float64, torch.float32],
     max_norm=[1.0, 10.0],
     lr=[1e-3, 1e-4],
     momentum=[0.0, 0.1],
